feat-npm scripts

+++ scripts/seeding.js
@@ -9,6 +9,7 @@ console.log('Entering seeding.js **********');
 
 const seedingDb = async (req, res) => {
   try {
+    if (process.env.NODE_ENV === 'production') throw new Error('You are not allowed to Seed Production Environment -do: >npm run dev')
     const clear = await db.sequelize.sync({ force: true });
     console.log('Clearing Data...')
     console.log('Creating DATA...')
@@ -44,7 +45,8 @@ const seedingDb = async (req, res) => {
     res.status(200).send('Seeds Created')
   } catch (err) {
     console.log(err);
-    res.status(500).send('Seeds NOT Created')
+    console.log('Seeds NOT Created')    
+    res.status(400).send(err.message)
   }
 }
